Rewrite with hciattach
Although btattach is more up to date, it doesn't permit setting the MAC
address of the adapter at initialization with the result that the MAC
was always AA:AA:AA:AA:AA:AA.

Also bought the script up to date with the current classic package:
support the CM4 with a wider match against the d-t model field, and
correct the UART for the mini-UART case (when serial1==uart0 in the
d-t). The gadget will need updating to support this (bt-serial needs to
include ttyS0).