Re-wrote the script to support Core
The bthelper script previously needed calling with the hci device name,
but this was managed by custom udev rules on classic which the snap
can't include. Instead, iterate over all the extant hci devices and
attempt to activate the first one based on a UART.

Also bought the script up to date with the check for the MAC and the
inclusion of the bluetoothctl reset although that is currently commented
out as I can't get bluetoothctl working within the snap (libreadline
conflict due to differing core bases).